+++ src/core/mem.h
@@ -6,7 +6,7 @@
 
 void *THE_PersistentAlloc(size_t size, size_t align);
 void *THE_Alloc(size_t size);
-void *THE_Calloc(int32_t elem_count, size_t elem_size);
+void *THE_Calloc(size_t elem_count, size_t elem_size);
 void *THE_Realloc(void *ptr, size_t size);
 void THE_Free(void *ptr);
